+++ mydatapreprocessing/load_data.py
@@ -19,6 +19,7 @@ import mylogging
 
 import importlib
 from pathlib import Path
+import io
 
 import pandas as pd
 
@@ -63,13 +64,13 @@ def get_file_paths(
 
 def load_data(
     data,
-    header=None,
-    csv_style=None,
+    header="infer",
+    csv_style="infer",
     predicted_table="",
     max_imported_length=0,
     request_datatype_suffix="",
     data_orientation="",
-    ssl_verification=False,
+    ssl_verification=None,
 ):
 
     """Load data from path or url or other python format (numpy array, list, dict) into dataframe.
@@ -81,9 +82,9 @@ def load_data(
 
     Args:
         data (Any): Path, url. For examples check examples section.
-        header (int, optional): Row index used as column names. Defaults to None.
-        csv_style ((dict, None), optional): Define CSV separator and decimal. En locale usually use {'sep': ",", 'decimal': "."}
-            some Europian country use {'sep': ";", 'decimal': ","}. If None - values are infered. Defaults to None.
+        header ((str, int), optional): Row index used as column names. If 'infer', it will be automatically choosed. Defaults to 'infer'.
+        csv_style ((dict, str), optional): Define CSV separator and decimal. En locale usually use {'sep': ",", 'decimal': "."}
+            some Europian country use {'sep': ";", 'decimal': ","}. If 'infer' one of those two locales is automatically used. Defaults to 'infer'.
         predicted_table (str, optional): If using excel (xlsx) - it means what sheet to use, if json,
             it means what key values, if SQL, then it mean what table. Else it have no impact. Defaults to ''.
         max_imported_length (int, optional): Max length of imported samples (before resampling). If 0, than full length.
@@ -92,7 +93,7 @@ def load_data(
             define whichdatatype is on this url with GET request. Defaults to "".
         data_orientation(str, optional): 'columns' or 'index'. If using json or dictionary, it describe how data are
             oriented. Default is 'columns' if None used. If orientation is records (in pandas terminology), it's detected
-            automatically. Defaults to "".
+            automatically (therefore empty by default). Defaults to "".
         ssl_verification(bool, optional): If using data from web, it use requests and sometimes, there can be ssl verification
             error, this skip verification, with adding verify param to requests call.
 
@@ -106,7 +107,7 @@ def load_data(
     Examples:
         You can use local files as well as web urls
 
-        >>> data_loaded = load_data("https://blockchain.info/unconfirmed-transactions?format=json", request_datatype_suffix=".json", data_orientation="index", predicted_table='txs')
+        >>> data_loaded = load_data("https://www.ncdc.noaa.gov/cag/global/time-series/globe/land_ocean/ytd/12/1880-2016.json", request_datatype_suffix=".json", predicted_table='data', data_orientation="index")
         >>> # data2 = load_data(PATH_TO_FILE.csv)
 
         Allowed data formats for load_data are examples::
@@ -114,8 +115,8 @@ def load_data(
             myarray_or_dataframe # Numpy array or Pandas.DataFrame
             r"/home/user/my.json" # Local file. The same with .parquet, .h5, .json or .xlsx.
             "https://yoururl/your.csv" # Web url (with suffix). Same with json.
-            "https://blockchain.info/unconfirmed-transactions?format=json" # In this case you have to specify
-                also 'request_datatype_suffix': "json", 'data_orientation': "index", 'predicted_table': 'txs',
+            "https://www.ncdc.noaa.gov/cag/global/time-series/globe/land_ocean/ytd/12/1880-2016.json" # In this case you have to specify
+                also 'request_datatype_suffix': "json", 'data_orientation': "index", 'predicted_table': 'data',
             [{'col_1': 3, 'col_2': 'a'}, {'col_1': 0, 'col_2': 'd'}] # List of records
             {'col_1': [3, 2, 1, 0], 'col_2': ['a', 'b', 'c', 'd']} # Dict with colums or rows (index) - necessary
                 to setup data_orientation!
@@ -128,9 +129,23 @@ def load_data(
                 "https://raw.githubusercontent.com/jbrownlee/Datasets/master/daily-min-temperatures.csv"]  # List of URLs
             ["path/to/my1.csv", "path/to/my1.csv"]
 
+        You can use some test data. Use one of 'test_ecg' 'test_ramp' 'test_sin' 'test_random'.
+
     Note:
         On windows it's necessary to use raw string - 'r' in front of string because of escape symbols \
     """
+    if isinstance(data, str):
+        if data == "test_ramp":
+            return generate_data.ramp()
+
+        elif data == "test_sin":
+            return generate_data.sin()
+
+        elif data == "test_random":
+            return generate_data.random()
+
+        elif data == "test_ecg":
+            return generate_data.get_ecg()
 
     list_of_dataframes = []
 
@@ -141,12 +156,34 @@ def load_data(
     if isinstance(data[0], (list, tuple)):
         return pd.DataFrame.from_records(data)
 
-    # If data is only path or URL or test or SQL
+    # If data is only path or URL
     elif isinstance(data, (list, tuple)) and isinstance(data[0], (str, Path)):
 
         for i in data:
             data_path = Path(i)
             iterated_data = None
+
+            # For example csv or json. On url, take everything after last dot
+            data_type_suffix = data_path.suffix[1:].lower()
+
+            # If not suffix inferred, then maybe url that return as request - than suffix have to be configured
+            if not data_type_suffix or (
+                data_type_suffix not in ["csv", "json", "xlsx", "xls"] and request_datatype_suffix
+            ):
+                data_type_suffix = request_datatype_suffix.lower()
+
+                if data_type_suffix.startswith("."):
+                    data_type_suffix = data_type_suffix[1:]
+
+            if not data_type_suffix:
+                raise TypeError(
+                    mylogging.return_str(
+                        "Data has no suffix (e.g. csv). If using url with no suffix, setup"
+                        "'request_datatype_suffix' or insert data with local path or insert data for example in"
+                        f"dataframe or numpy array. \n\nYour configured data are {data}",
+                        caption="Data load error",
+                    )
+                )
 
             try:
                 if data_path.exists():
@@ -179,54 +216,43 @@ def load_data(
                         )
                     )
 
-            # For example csv or json. On url, take everything after last dot
-            data_type_suffix = data_path.suffix[1:].lower()
-
-            # If not suffix inferred, then maybe url that return as request - than suffix have to be configured
-            if not data_type_suffix or (
-                data_type_suffix not in ["csv", "json", "xlsx", "xls"] and request_datatype_suffix
-            ):
-                data_type_suffix = request_datatype_suffix.lower()
-
-                if data_type_suffix.startswith("."):
-                    data_type_suffix = data_type_suffix[1:]
-
-            # If it's URL with suffix, we usually need url, if its url link with no suffix, we need get request response
-            if not is_file:
                 if data_type_suffix == "json":
                     iterated_data = request.content
 
                 if data_type_suffix == "csv":
-                    iterated_data = i
-
-            if not data_type_suffix:
-                raise TypeError(
-                    mylogging.return_str(
-                        "Data has no suffix (e.g. csv). If using url with no suffix, setup"
-                        "'request_datatype_suffix' or insert data with local path or insert data for example in"
-                        f"dataframe or numpy array. \n\nYour configured data are {data}",
-                        caption="Data load error",
-                    )
-                )
+                    iterated_data = io.BytesIO(request.content)
 
             if data_type_suffix == "csv":
 
-                if not header:
-                    header = "infer"
+                if csv_style == "infer":
+                    # Skip some lines if comments there
+                    if isinstance(iterated_data, io.BytesIO):
+                        data_line = iterated_data.readline()
+                        for i in range(20):
+                            new_line = iterated_data.readline()
+                            if new_line:
+                                data_line = new_line
+                            else:
+                                break
+                        iterated_data.seek(0)
 
-                if not csv_style:
-                    sep = pd.read_csv(iterated_data, sep=None, iterator=True)._engine.data.dialect.delimiter
+                    else:
+                        with open(iterated_data, "r") as data_read:
+                            data_line = data_read.readline()
+                            for i in range(20):
+                                new_line = data_read.readline()
+                                if new_line:
+                                    data_line = new_line
+                                else:
+                                    break
 
-                    if sep not in [",", ";", "\t"]:
-                        raise ValueError(
-                            mylogging.return_str(
-                                "CSV separator not infered. Infering not possible if description with symbols on "
-                                "first few lines. Define parameter csv_style - separator and decimal manually and "
-                                "skip description with header parameter."
-                            )
-                        )
-
-                    decimal = "," if sep == ";" else "."
+                    data_line = str(data_line)
+                    if data_line.count(";") and data_line.count(";") >= data_line.count(",") - 1:
+                        sep = ";"
+                        decimal = ","
+                    else:
+                        sep = ","
+                        decimal = "."
 
                 else:
                     sep = csv_style["sep"]
@@ -243,10 +269,14 @@ def load_data(
                         pd.read_csv(
                             iterated_data,
                             header=header,
-                            sep=csv_style["sep"],
-                            decimal=csv_style["decimal"],
+                            sep=sep,
+                            decimal=decimal,
                             encoding="cp1252",
                         ).iloc[-max_imported_length:, :]
+                    )
+                except Exception:
+                    raise RuntimeError(
+                        mylogging.return_str("CSV load failed. Try to set correct `header` and `csv_style`")
                     )
 
             elif data_type_suffix == "xls":