+++ changelog.md
@@ -5,6 +5,7 @@
 New formats support. H5, Parquet, xlsx...
 Multiple files at once in a list ['file1.csv', 'file2.csv']
 New function get_file_paths that open dialog window and let you choose files very simply. Return list of paths.
+String embedding - transform string columns - E.g. Country ('US', 'FR'... etc.) into numbers so it can be used in machine learning models.
 Deriving new columns with rolling fast fourier transform in preprocessing with add_frequency_columns()
 Deriving new columns with add_derived_columns() in preprocessing such as first and second difference or rolling means and rolling standard deviations
 Inputs model described in readme